ENH: Expand the flexibility of the constraints parser
